pyproject.toml
[tool.poetry]
name = "edvardks-portfolio"
version = "0.2.0"
description = "Personal site (Django + Tailwind, modular per-component)."
authors = ["Edvard Khachatryan Sahakyan <info@edvardks.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
